Move notify_action files into actions/ modules
