+++ src/sentry/integrations/jira/__init__.py
@@ -1,8 +1,8 @@
 from sentry.rules import rules
 
+from .actions import JiraCreateTicketAction
 from .client import JIRA_KEY, JiraApiClient
 from .integration import JiraIntegration, JiraIntegrationProvider
-from .notify_action import JiraCreateTicketAction
 
 __all__ = (
     "JIRA_KEY",
