+++ src/sentry/integrations/jira/actions/__init__.py
@@ -1,0 +1,7 @@
+from .create_ticket import JiraCreateTicketAction
+from .form import JiraNotifyServiceForm
+
+__all__ = (
+    "JiraCreateTicketAction",
+    "JiraNotifyServiceForm",
+)
