+++ src/sentry/integrations/jira/actions/create_ticket.py
@@ -1,28 +1,11 @@
 from __future__ import annotations
 
-import logging
 from typing import Any
 
-from django import forms
-from django.utils.translation import ugettext_lazy as _
-
+from sentry.integrations.jira.actions.form import JiraNotifyServiceForm
 from sentry.models import Integration
-from sentry.rules.actions import IntegrationNotifyServiceForm, TicketEventAction
+from sentry.rules.actions import TicketEventAction
 from sentry.utils.http import absolute_uri
-
-logger = logging.getLogger("sentry.rules")
-
-
-class JiraNotifyServiceForm(IntegrationNotifyServiceForm):
-    def clean(self) -> dict[str, Any] | None:
-        cleaned_data = super().clean()
-
-        integration = cleaned_data.get("integration")
-        try:
-            Integration.objects.get(id=integration)
-        except Integration.DoesNotExist:
-            raise forms.ValidationError(_("Jira integration is a required field."), code="invalid")
-        return cleaned_data
 
 
 class JiraCreateTicketAction(TicketEventAction):
