+++ src/sentry/integrations/msteams/__init__.py
@@ -1,7 +1,7 @@
 from sentry.rules import rules
 from sentry.utils.imports import import_submodules
 
-from .notify_action import MsTeamsNotifyServiceAction
+from .actions import MsTeamsNotifyServiceAction
 
 import_submodules(globals(), __name__, __path__)
