+++ src/sentry/integrations/msteams/actions/__init__.py
@@ -1,0 +1,7 @@
+from .form import MsTeamsNotifyServiceForm
+from .notification import MsTeamsNotifyServiceAction
+
+__all__ = (
+    "MsTeamsNotifyServiceForm",
+    "MsTeamsNotifyServiceAction",
+)
