+++ src/sentry/integrations/pagerduty/actions/__init__.py
@@ -1,0 +1,7 @@
+from .form import PagerDutyNotifyServiceForm
+from .notification import PagerDutyNotifyServiceAction
+
+__all__ = (
+    "PagerDutyNotifyServiceForm",
+    "PagerDutyNotifyServiceAction",
+)
