+++ src/sentry/integrations/slack/actions/__init__.py
@@ -1,0 +1,7 @@
+from .form import SlackNotifyServiceForm
+from .notification import SlackNotifyServiceAction
+
+__all__ = (
+    "SlackNotifyServiceForm",
+    "SlackNotifyServiceAction",
+)
