+++ src/sentry/integrations/vsts/__init__.py
@@ -1,7 +1,7 @@
 from sentry.rules import rules
 from sentry.utils.imports import import_submodules
 
-from .notify_action import AzureDevopsCreateTicketAction
+from .actions import AzureDevopsCreateTicketAction
 
 rules.add(AzureDevopsCreateTicketAction)
 import_submodules(globals(), __name__, __path__)  # type: ignore
