+++ src/sentry/integrations/vsts/actions/__init__.py
@@ -1,0 +1,3 @@
+from .create_ticket import AzureDevopsCreateTicketAction
+
+__all__ = ("AzureDevopsCreateTicketAction",)
