+++ tests/sentry/integrations/jira/test_notify_action.py
@@ -1,6 +1,6 @@
 import responses
 
-from sentry.integrations.jira.notify_action import JiraCreateTicketAction
+from sentry.integrations.jira import JiraCreateTicketAction
 from sentry.models import ExternalIssue, GroupLink, Integration, Rule
 from sentry.testutils.cases import RuleTestCase
 from sentry.types.rules import RuleFuture
